+++ controllers/protocols/http/host.js
@@ -11,6 +11,7 @@
     const Error         = require('../../error')
     const url           = require('url')
     const RouteIterator = require('../../iterator')
+    const utils         = require('../../utils')
     
 // ===========================================================================
 //  Exports
@@ -30,7 +31,7 @@
     		var host     = request.headers.host ? request.headers.host : '' ; 
     		var location = request.url ? url.parse(request.url) : '' ;      
     		var port     = host.split(':')[1] || protocol.globalAgent.defaultPort;
-    		var hostname = isset(port) ? host : host + ':' + protocol.globalAgent.defaultPort ;
+    		var hostname = utils.isset(port) ? host : host + ':' + protocol.globalAgent.defaultPort ;
 			
     		// get app (host controller) handling this hostname
     		var app = App.hosts[hostname] || App.hosts[hostname+':'+port] || App.hosts['0.0.0.0:'+port]