+++ controllers/utils.js
@@ -1,0 +1,3 @@
+module.exports.isset = function isset(object){
+	return (object != "undefined" && object != undefined && object != null && object != "" && typeof(object) != 'undefined') ? true : false ;
+}